real_estate_inquiries_agent/requirements.txt
azure-ai-agentserver-agentframework==1.0.0b16
azure-ai-agentserver-core==1.0.0b16
agent-dev-cli==0.0.1b260316
agent-framework-core==1.0.0rc6
agent-framework-foundry==1.0.0rc6
azure-identity==1.15.0
python-dotenv==1.0.0
requests
uvicorn
aiohttp
debugpy